Clase 08/requirements.txt
python-dotenv
httpx
openai-agents
sendgrid==6.11.0
ipykernel
gradio
